* gm2/Make-lang.in:  added extra commented definition for bootstrapping under trunc. * gm2/bnf/mc-1.bnf:  changed comment interface. * gm2/bnf/mc-2.bnf:  removed unnecessary import. * gm2/bnf/mc-5.bnf:  rewritten the comment interface for   after and body comments. * gm2/mc-boot/GCmdArgs.c:  rebuilt. * gm2/mc-boot/GDebug.c:  rebuilt. * gm2/mc-boot/GDynamicStrings.c:  rebuilt. * gm2/mc-boot/GFIO.c:  rebuilt. * gm2/mc-boot/GFormatStrings.c:  rebuilt. * gm2/mc-boot/GFpuIO.c:  rebuilt. * gm2/mc-boot/GIO.c:  rebuilt. * gm2/mc-boot/GIndexing.c:  rebuilt. * gm2/mc-boot/GM2EXCEPTION.c:  rebuilt. * gm2/mc-boot/GM2RTS.c:  rebuilt. * gm2/mc-boot/GM2RTS.h:  rebuilt. * gm2/mc-boot/GNumberIO.c:  rebuilt. * gm2/mc-boot/GRTExceptions.c:  rebuilt. * gm2/mc-boot/GRTint.c:  rebuilt. * gm2/mc-boot/GSFIO.c:  rebuilt. * gm2/mc-boot/GStorage.c:  rebuilt. * gm2/mc-boot/GStrIO.c:  rebuilt. * gm2/mc-boot/GStrLib.c:  rebuilt. * gm2/mc-boot/GStringConvert.c:  rebuilt. * gm2/mc-boot/GSysStorage.c:  rebuilt. * gm2/mc-boot/Gdecl.c:  rebuilt. * gm2/mc-boot/Gdecl.h:  rebuilt. * gm2/mc-boot/Gkeyc.c:  rebuilt. * gm2/mc-boot/Glibc.h:  rebuilt. * gm2/mc-boot/GmcComment.c:  rebuilt. * gm2/mc-boot/GmcComment.h:  rebuilt. * gm2/mc-boot/GmcComp.c:  rebuilt. * gm2/mc-boot/GmcComp.h:  rebuilt. * gm2/mc-boot/GmcError.c:  rebuilt. * gm2/mc-boot/GmcLexBuf.c:  rebuilt. * gm2/mc-boot/GmcLexBuf.h:  rebuilt. * gm2/mc-boot/GmcMetaError.c:  rebuilt. * gm2/mc-boot/GmcOptions.c:  rebuilt. * gm2/mc-boot/GmcPretty.c:  rebuilt. * gm2/mc-boot/GmcPrintf.c:  rebuilt. * gm2/mc-boot/GmcQuiet.c:  rebuilt. * gm2/mc-boot/GmcSearch.c:  rebuilt. * gm2/mc-boot/Gmcp1.c:  rebuilt. * gm2/mc-boot/Gmcp2.c:  rebuilt. * gm2/mc-boot/Gmcp3.c:  rebuilt. * gm2/mc-boot/Gmcp4.c:  rebuilt. * gm2/mc-boot/Gmcp5.c:  rebuilt. * gm2/mc-boot/GnameKey.c:  rebuilt. * gm2/mc-boot/GsymbolKey.c:  rebuilt. * gm2/mc-boot/Gvarargs.c:  rebuilt. * gm2/mc-boot/Gwrapc.h:  rebuilt. * gm2/mc/decl.def:  (addCommentBody)  New procedure.   (addCommentAfter)  New procedure.  (addIfComments)   New procedure.  (makeCommentS)  New procedure.   (getCommentCharStar) New procedure function.   (getContent) New procedure function.   (setProcedureComment) New procedure function.   (getProcedureComment) New procedure function.   (getAfterStatementComment)  New procedure function.   (getInbodyStatementComment)  New procedure function.   (isProcedureComment)  New procedure function.   (isBodyComment)  New procedure function.   (isAfterComment)  New procedure function. * gm2/mc/decl.mod:  many alterations to allow per node   after and body comments. * gm2/mc/keyc.mod:  add free and malloc to the list of   names to avoid. * gm2/mc/mc.flex:  many alterations to allow per node   after and body comments. * gm2/mc/mcComment.def:  (commentDesc)  New datatype.   (initComment)  New procedure function.  (addText)   New procedure.  (getContent)  New procedure function. * gm2/mc/mcComment.h:  updated with new interface. * gm2/mc/mcComment.mod:  (commentDesc)  New datatype   implemented.   (initComment)  New procedure function.  (addText)   New procedure.  (getContent)  New procedure function. * gm2/mc/mcComp.def:  export getPassNo. * gm2/mc/mcComp.mod:  (getPassNo)  New procedure function. * gm2/mc/mcLexBuf.def:  (getProcedureComment)  New procedure   function.  (getBodyComment)  New procedure function.   (getAfterComment)  New procedure function. * gm2/mc/mcLexBuf.h:  replicate .def functionality. * gm2/mc/mcLexBuf.mod:  (getProcedureComment)  Implement new   procedure function.  (getBodyComment)  Implement new   procedure function.

+++ gcc-versionno/gcc/gm2/mc-boot/GIO.c
@@ -14,6 +14,7 @@
 #      define FALSE (1==0)
 #   endif
 
+#include <stdlib.h>
 #define _IO_H
 #define _IO_C
 
@@ -143,7 +144,7 @@ static void doWrite (int fd, FIO_File f, char ch)
       if (! fdState.array[fd].IsEof)
         for (;;)
         {
-          r = libc_write (FIO_GetUnixFileDescriptor (f), &ch, 1);
+          r = libc_write (FIO_GetUnixFileDescriptor (f), &ch, (size_t) 1);
           if (r == 1)
             return;
           else if (r == -1)
@@ -180,6 +181,15 @@ static void setFlag (termios_TERMIOS t, termios_Flag f, unsigned int b)
 
 static void doraw (termios_TERMIOS term)
 {
+  /* 
+    * from man 3 termios
+    *           termios_p->c_iflag &= ~(IGNBRK | BRKINT | PARMRK | ISTRIP
+    *                                   | INLCR | IGNCR | ICRNL | IXON);
+    *           termios_p->c_oflag &= ~OPOST;
+    *           termios_p->c_lflag &= ~(ECHO | ECHONL | ICANON | ISIG | IEXTEN);
+    *           termios_p->c_cflag &= ~(CSIZE | PARENB);
+    *           termios_p->c_cflag |= CS8;
+  */
   setFlag (term, (termios_Flag) termios_ignbrk, FALSE);
   setFlag (term, (termios_Flag) termios_ibrkint, FALSE);
   setFlag (term, (termios_Flag) termios_iparmrk, FALSE);
@@ -206,6 +216,17 @@ static void doraw (termios_TERMIOS term)
 
 static void dononraw (termios_TERMIOS term)
 {
+  /* 
+    * we undo these settings, (although we leave the character size alone)
+    *
+    * from man 3 termios
+    *           termios_p->c_iflag &= ~(IGNBRK | BRKINT | PARMRK | ISTRIP
+    *                                   | INLCR | IGNCR | ICRNL | IXON);
+    *           termios_p->c_oflag &= ~OPOST;
+    *           termios_p->c_lflag &= ~(ECHO | ECHONL | ICANON | ISIG | IEXTEN);
+    *           termios_p->c_cflag &= ~(CSIZE | PARENB);
+    *           termios_p->c_cflag |= CS8;
+  */
   setFlag (term, (termios_Flag) termios_ignbrk, TRUE);
   setFlag (term, (termios_Flag) termios_ibrkint, TRUE);
   setFlag (term, (termios_Flag) termios_iparmrk, TRUE);
@@ -254,7 +275,7 @@ void IO_Read (char *ch)
     else
       for (;;)
       {
-        r = libc_read (FIO_GetUnixFileDescriptor (FIO_StdIn), ch, 1);
+        r = libc_read (FIO_GetUnixFileDescriptor (FIO_StdIn), ch, (size_t) 1);
         if (r == 1)
           return;
         else if (r == -1)
